Needed changes to make this compile after moving sensor.java.
