+++ app/src/main/java/com/eveningoutpost/dexdrip/Models/Sensor.java
@@ -1,4 +1,4 @@
-package com.eveningoutpost.dexdrip;
+package com.eveningoutpost.dexdrip.Models;
 
 import android.provider.BaseColumns;
